% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_sig.R
\name{beta_sig}
\alias{beta_sig}
\title{Compute the β (beta) parameter for the SiGN model}
\usage{
beta_sig(
  use_beta,
  sig,
  sched1,
  sched2,
  il_dur1,
  il_dur2,
  s_plus_dur,
  base = 10
)
}
\arguments{
\item{use_beta}{Logical vector indicating whether the β parameter
should be applied.}

\item{sig}{Logical vector indicating whether the alternative is signalled.
See \link{sig_check}.}

\item{sched1}{Character vector indicating the schedule type for the first
alternative's initial link. See Details.
(e.g., \code{"VI"}, \code{"FR"}).}

\item{sched2}{Character vector indicating the schedule type for the second
alternative's initial link. See Details.}

\item{il_dur1}{Numeric vector of durations for the first alternative's
initial link.}

\item{il_dur2}{Numeric vector of durations for the second alternative's
initial link.}

\item{s_plus_dur}{Numeric vector of durations associated with alternative
1's \eqn{S^{+}} stimulus.}

\item{base}{A positive real number specifying the base of
the logarithm used for β. Defaults to 10.}
}
\value{
A numeric vector containing the β values.
}
\description{
Computes the β adjustment for signalled alternatives, reflecting the
balance between conditional and terminal (i.e., primary/unconditional)
reinforcement, as described in Equation 6 of Dunn et al. (2024).
}
\details{
Although the initial link durations for both alternatives are provided as
arguments, the function computes β specifically for the first initial
link.

If both initial links are variable interval (VI) schedules, a common initial
link duration is computed: \deqn{1 / (1 / d_1 + 1 / d_2)}
Otherwise, the duration of the first initial link is used.

See \code{\link[=choice_params]{choice_params()}} for additional details appropriate on
schedule selection.
}
\examples{
beta_sig(
  use_beta = TRUE,
  sig = TRUE,
  sched1 = "FR",
  sched2 = "FR",
  il_dur1 = 1,
  il_dur2 = 1,
  s_plus_dur = 10
)
}
\references{
Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2024).
Suboptimal choice: A review and quantification of the signal for good news
(SiGN) model. \emph{Psychological Review}. \emph{131}(1), 58-78.
\url{https://doi.org/10.1037/rev0000416}
}
