% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleSort.R
\name{BubbleSort}
\alias{BubbleSort}
\title{BubbleSort is a function for calculating bubble sort correlation between two vectors}
\usage{
BubbleSort(Vec1, Vec2)
}
\arguments{
\item{Vec1}{Vector of values of 1st feature across samples}

\item{Vec2}{Vector of values of 2nd feature across samples}
}
\value{
Bubble sort similarity between the two vectors
}
\description{
BubbleSort is a function for calculating bubble sort correlation between two vectors
}
