% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_2019.R
\docType{data}
\name{cs_2019}
\alias{cs_2019}
\title{Suboptimal Choice Data from Cunningham & Shahan (2019)}
\format{
A data frame with 50 rows and 22 columns:
\describe{
\item{subject}{A unique identifier for each rat.}
\item{cp}{Mean suboptimal choice proportion.}
\item{il_dur_a, il_dur_b}{Programmed initial link (IL) durations
for alternatives A and B, respectively. FR1 schedules are assigned a
duration of 1.}
\item{tl_dur_a1, tl_dur_a2, tl_dur_b1, tl_dur_b2}{Programmed
durations of terminal links. For example, \code{tl_dur_a1} corresponds to the
first terminal link on alternative A.}
\item{tl_p_a1, tl_p_a2, tl_p_b1, tl_p_b2}{Probability of entering each
respective terminal link.}
\item{tr_p_a1, tr_p_a2, tr_p_b1, tr_p_b2}{Probability of terminal (i.e.,
primary) reinforcement following entry into each respective terminal link.}
\item{il_sched_a, il_sched_b}{Schedule type used for the initial links.}
\item{tl_sched_a1, tl_sched_a2, tl_sched_b1, tl_sched_b2}{Schedule type
used for each terminal link.}
}
}
\source{
Cunningham, P. J., & Shahan, T. A. (2019). Rats engage in suboptimal choice
when the delay to food is sufficiently long. \emph{Journal of Experimental
Psychology. Animal Learning and Cognition}, \emph{45}(3), 301–310.
https://doi.org/10.1037/xan0000211
}
\usage{
cs_2019
}
\description{
Suboptimal choice data from Cunningham and Shahan's (2019) study, \emph{Rats
Engage in Suboptimal Choice When the Delay to Food Is Sufficiently Long}.
The dataset includes choice behaviour observed in rats under varying
terminal link delay conditions.
}
\details{
The data were manually extracted from published figures using the
\pkg{digitize} package. While care was taken to ensure accuracy, slight
deviations from the original values are possible due to limitations in
figure resolution and interpretation.
}
\keyword{datasets}
