% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_bonus.R
\name{dr_bonus}
\alias{dr_bonus}
\title{Bonus Delay Reduction}
\usage{
dr_bonus(sig, big_t, s_plus_tr_p, s_plus_dur, tr_p, tl_dur)
}
\arguments{
\item{sig}{Logical vector indicating whether the choice alternative is
signalled.}

\item{big_t}{Numeric vector of the overall average programmed time to
terminal reinforcement from the start of the initial links (excluding
time spent in a intertrial interval). See \code{\link[=SiGN]{SiGN()}}.}

\item{s_plus_tr_p}{Numeric vector of terminal reinforcement probabilities
for an alternative's \eqn{S^{+}} terminal link. See \code{\link[=s_plus_tr_p]{s_plus_tr_p()}}.}

\item{s_plus_dur}{Numeric vector of durations associated with an
alternative's \eqn{S^{+}} terminal link. See \code{\link[=s_plus_dur]{s_plus_dur()}}.}

\item{tr_p}{Numeric vector of overall terminal reinforcement probabilities
for the alternative.}

\item{tl_dur}{Numeric vector of average terminal link durations for the
alternative.}
}
\value{
A numeric vector representing the bonus delay reduction.
}
\description{
A helper function to compute bonus delay reduction on a alternative.
}
\details{
If the alternative is not signalled (\code{sig == FALSE}), the bonus is set to 0.
}
\examples{
dr_bonus(
  sig = TRUE,
  big_t = 20,
  s_plus_tr_p = 1,
  s_plus_dur = 10,
  tr_p = 0.2,
  tl_dur = 10
)
}
