% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_p_sums.R
\name{entry_p_sums}
\alias{entry_p_sums}
\title{Sum Entry Probabilities Across Four Terminal Links}
\usage{
entry_p_sums(x)
}
\arguments{
\item{x}{A list of 4 numeric vectors with an equal number of elements.}
}
\value{
A numeric vector of sums, where each element corresponds to
the sum of the four probabilities.
}
\description{
This function takes a list of 4 numeric vectors (each representing entry
probabilities for a terminal link) and returns the sum of the probabilities
across the 4 vectors.
}
\examples{
entry_p_sums(list(
  a = c(0.2, 0.5, 1),
  b = c(0.8, 0.5, 0),
  c = c(0.2, 0.5, 1),
  d = c(0.8, 0.5, 1)
))
}
