% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventRenaming.R
\name{EventRenaming}
\alias{EventRenaming}
\title{EventRenaming is a function for changing annotation of censored samples to 0
 and dead samples to 1 for survival analysis}
\usage{
EventRenaming(EventVec, Censored_Annot)
}
\arguments{
\item{EventVec}{Status vector for all of the samples (patients) including both samples undergone an event or censored}

\item{Censored_Annot}{Index of samples censored in the dataset}
}
\value{
Vector of events including 0 for censoring and 1 for death
}
\description{
EventRenaming is a function for changing annotation of censored samples to 0
 and dead samples to 1 for survival analysis
}
