% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpPhen_Matching.R
\name{ExpPhen_Matching}
\alias{ExpPhen_Matching}
\title{ExpPhen_Matching is a function for matching samples between expression matrices 
and metadata matrix (clinical feature matrix)}
\usage{
ExpPhen_Matching(ExpMat, MetaMat, SamID_Meta)
}
\arguments{
\item{ExpMat}{Matrix of expression of genes (samples in columns and genes in rows)}

\item{MetaMat}{Matrix of clinical features (samples in columns)}

\item{SamID_Meta}{Sample ID in MetaMat}
}
\value{
List of expression matrix and metadata of the clinical information after 
matching patiend IDs between the expression and clinical information matrices
}
\description{
ExpPhen_Matching is a function for matching samples between expression matrices 
and metadata matrix (clinical feature matrix)
}
