% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpPhen_Subdividing.R
\name{ExpPhen_Subdividing}
\alias{ExpPhen_Subdividing}
\title{ExpPhen_Subdividing is a function for grouping samples based on a clinical feature
 available in metadata matrix (clinical feature matrix)}
\usage{
ExpPhen_Subdividing(ExpMeta_List, SubDiv_ID)
}
\arguments{
\item{ExpMeta_List}{List containing expression matrix and metadata matrix}

\item{SubDiv_ID}{Index of the target clinical feature in metadata matrix for samples grouping}
}
\value{
List of expression and clinical information of patients grouped based on the specified clinical feature
}
\description{
ExpPhen_Subdividing is a function for grouping samples based on a clinical feature
 available in metadata matrix (clinical feature matrix)
}
