% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpPheno_Categorize.R
\name{ExpPheno_Categorize}
\alias{ExpPheno_Categorize}
\title{ExpPheno_Categorize is a function for grouping samples 
based on their survival to 3 groups of poor, good, and intermediate}
\usage{
ExpPheno_Categorize(ExpMeta_List, Time_ID, Event_ID, Mad_Factor,
  MinNum_ExClass, Expression_Log2 = FALSE)
}
\arguments{
\item{ExpMeta_List}{List containing expression matrix and metadata matrix}

\item{Time_ID}{Index of time to death in metadata matrix}

\item{Event_ID}{Index of event in metadata matrix}

\item{Mad_Factor}{Threshold of mad in time to death values to determine poor survival group}

\item{MinNum_ExClass}{Minimum number of samples that has to be kept in poor and good group 
(if number of samples is lower than this threhold, more samples will be addedd in order of survival)}

\item{Expression_Log2}{Parameter for gene expsression value transformation to logarithmic scale (log2(expression value+1))}
}
\value{
List of expression matrices, and time to event as well as event for the patients
 within each category of poor, intermediate or good survival
}
\description{
ExpPheno_Categorize is a function for grouping samples 
based on their survival to 3 groups of poor, good, and intermediate
}
