% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneMatching.R
\name{GeneMatching}
\alias{GeneMatching}
\title{GeneMatching is a function to remove uncommon genes between a list of expression matrices}
\usage{
GeneMatching(ExpList)
}
\arguments{
\item{ExpList}{List of expression matrices}
}
\value{
List of expression matrices restricted to the common genes between them
}
\description{
GeneMatching is a function to remove uncommon genes between a list of expression matrices
}
