% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Grouping.R
\name{Pathway_Grouping}
\alias{Pathway_Grouping}
\title{Pathway_Grouping is a function to make a pathway list from files containing genes within each pathway}
\usage{
Pathway_Grouping(PathwayDir, Pattern)
}
\arguments{
\item{PathwayDir}{Path of directory including the files of pathways}

\item{Pattern}{Pattern should be used to select the files of pathway genes from PathwayDir}
}
\value{
List of genes within the pathway
}
\description{
Pathway_Grouping is a function to make a pathway list from files containing genes within each pathway
}
