% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Similarity.R
\name{Pathway_similarity}
\alias{Pathway_similarity}
\title{Pathway_similarity is a function for calculating correlation between expression level of pathways
 between 2 groups using all the available approaches in SIGN}
\usage{
Pathway_similarity(ExpMat1, ExpMat2, GeneVec, GeneSets, Name)
}
\arguments{
\item{ExpMat1}{Expression matrix of genes in the 1st group of sampls}

\item{ExpMat2}{Expression matrix of genes in the 2nd group of sampls}

\item{GeneVec}{Name of genes in the same order as considered in ExpMat1 and ExpMat2}

\item{GeneSets}{List of genes within pathways}

\item{Name}{Name used for naming the columns of output matrix of correlation between the 2 groups}
}
\value{
Similarity of the pathway between the two expression matrices using 
pearson correlation, bubble sort, and wilcoxon paaired rank test
}
\description{
Pathway_similarity is a function for calculating correlation between expression level of pathways
 between 2 groups using all the available approaches in SIGN
}
