% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_SiGN.R
\name{pred_SiGN}
\alias{pred_SiGN}
\title{Compute SiGN Prediction}
\usage{
pred_SiGN(r1, r2, cr1, cr2)
}
\arguments{
\item{r1}{Numeric vector of terminal reinforcement rates for alternative 1.}

\item{r2}{Numeric vector of terminal reinforcement rates for alternative 2.}

\item{cr1}{Numeric vector of \eqn{\delta} values for alternative 1.}

\item{cr2}{Numeric vector of \eqn{\delta} values for alternative 2.

@details
The underlying calculation is described in Dunn et al. (2024), with
Equation 7 providing the core formulation.}
}
\value{
A numeric vector of predicted choice probabilities for alternative 1.
}
\description{
A helper function compute the probability of selecting the first alternative
according to the SiGN (Signal for Good News; Dunn et al. 2024) and by
extension DRH (Delay Reduction Hypothesis; Squires & Fantino, 1971).
}
\examples{
pred_SiGN(r1 = 0.05, r2 = 0.05, cr1 = 12, cr2 = 0.6)
pred_SiGN(r1 = 0.05, r2 = 0.01, cr1 = 30, cr2 = -30)  # returns 1 (edge case)
}
\references{
Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2024).
Suboptimal choice: A review and quantification of the signal for good news
(SiGN) model. \emph{Psychological Review}. \emph{131}(1), 58-78.
\url{https://doi.org/10.1037/rev0000416}

Squires, N., Fantino, E. (1971). A model for choice in simple concurrent and
concurrent-chains schedules. \emph{Journal of the Experimental Analysis of
Behavior}. \emph{15}(1), 27 - 38. \url{https://doi.org/10.1901/jeab.1971.15-27}
}
