% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rein_rate_function.R
\name{r}
\alias{r}
\title{Compute Rate of (Terminal) Reinforcement}
\usage{
r(
 il_dur1, il_dur2,
 tl_dur1, tr_p,
 il_sched1, il_sched2,
 common_il_dur = FALSE
 )
}
\arguments{
\item{il_dur1}{Duration of the first alternative's IL.}

\item{il_dur2}{Duration of the second alternative's IL.}

\item{tl_dur1}{Duration of the first alternative's TL.}

\item{tr_p}{Probability of terminal reinforcement.}

\item{il_sched1}{Schedule type for the initial links. (e.g., \code{"VI"}
or \code{"FR"}). See \code{\link[=choice_params]{choice_params()}} for additional details about inputting
schedule types.}

\item{il_sched2}{Schedule type for the second alternative's initial link.}

\item{common_il_dur}{Logical; if \code{TRUE}, a common initial link (IL) duration
is computed, but only when both IL schedules are \code{"VI"}. If \code{FALSE}, the
reinforcement rate is computed using only the first alternative's IL
duration, regardless of schedule type.}
}
\value{
A numeric value (or vector) giving the computed terminal
reinforcement rate.
}
\description{
Calculates the terminal reinforcement rate given initial link (IL)
durations, terminal link (TL) duration, terminal reinforcement probability,
and schedule types. Optionally computes a common effective IL
duration when both schedules are variable interval (VI).
}
\details{
When using vectorised inputs, arguments of differing lengths will be
recycled according to R's standard vector recycling rules. If lengths are
not compatible (i.e., not clean multiples), a warning may be issued.
See \code{\link[base:Arithmetic]{base::Arithmetic()}} for details.
}
\examples{
r(il_dur1 = 90, il_dur2 = 30, tl_dur1 = 30, tr_p = 1.0,
  il_sched1 = "VI", il_sched2 = "VI", common_il_dur = TRUE)

r(il_dur1 = 90, il_dur2 = NULL, tl_dur1 = 30, tr_p = 1.0,
  il_sched1 = "VI", il_sched2 = "VI", common_il_dur = FALSE)
}
\seealso{
\code{\link[=choice_params]{choice_params()}}
}
