% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_delta.R
\name{s_delta}
\alias{s_delta}
\title{Terminal Link Duration Adjustment for S- Signals}
\usage{
s_delta(tl_dur, tr_p, s_delta)
}
\arguments{
\item{tl_dur}{Numeric vector. Duration of the terminal link.}

\item{tr_p}{Numeric vector. Probability of terminal reinforcement following
the terminal link.}

\item{s_delta}{Numeric scalar or vector. Duration of the S- signal to use
when terminal reinforcement probability is zero.}
}
\value{
A numeric vector where elements of \code{tl_dur} are replaced by
\code{s_delta} if the terminal link has a non-zero duration and a
reinforcement probability of zero.
}
\description{
Adjusts the terminal link duration for signalled extinction.
}
\details{
\emph{Signal Discrimination:}

The \code{SiGN()} function assumes that terminal links which consistently lead to
extinction (i.e., no primary reinforcement) are fully discriminable by the
organism. In other words, the model treats terminal links that do and do not
lead to terminal reinforcement as completely distinct stimuli.

For example, suppose selecting Alternative A results in two equally probable
terminal links, both lasting 10 seconds—one always leads to reinforcement,
the other never does. The \code{SiGN()} function assumes these two links are
completely discriminable (e.g., by colour or some other cue).

If the organism cannot discriminate between them, this should be reflected
in how the parameters are specified. In such a case, rather than setting
\code{tr_p_a1 = 1} and \code{tr_p_a2 = 0}, you would model the ambiguity by setting
\code{tr_p_a1 = 0.5} and \code{tr_p_a2 = 0.5}.

\emph{Temporal Units:}

Conventionally, concurrent-chain paradigms have measured time in seconds.
The default value of \code{s_delta} is chosen on the basis of that convention.
However, given the temporal relatively of the SiGN model, if other units are
used, it may be prudent to adjust \code{s_delta} accordingly.
}
\examples{
s_delta(
  tl_dur = c(10, 10, 10, 0),
  tr_p = c(0, 1, 0.5, 1),
  s_delta = 1
  )
}
\references{
Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2023).
Suboptimal choice: A review and quantification of the signal for good news
(SiGN) model. \emph{Psychological Review}. \emph{131}(1), 58-78.
\url{https://doi.org/10.1037/rev0000416}
}
\seealso{
\code{\link[=choice_params]{choice_params()}}
}
