% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGN_Aggregate.R
\name{SIGN_Aggregate}
\alias{SIGN_Aggregate}
\title{SIGN_Aggregate is a function to reshape the list of pathway scoring, time to death, and event 
and return a summary list}
\usage{
SIGN_Aggregate(ScoreList, TimeList, EventList)
}
\arguments{
\item{ScoreList}{List of similarity scores identified using different methodologies}

\item{TimeList}{List of time to event (death) for different groups of patients}

\item{EventList}{List of event vectors (death or censored) for different groups of patients}
}
\value{
List of scores identified for each sample as well as time to death and event of that sample
}
\description{
SIGN_Aggregate is a function to reshape the list of pathway scoring, time to death, and event 
and return a summary list
}
