% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarities_Wrapper.R
\name{Similarities_Wrapper}
\alias{Similarities_Wrapper}
\title{Similarities_Wrapper is wrapper to identify similarities between the expression of genes in target sample
and the reference expression matrix}
\usage{
Similarities_Wrapper(ExpMat_Test, ExpMat_Ref, GeneVec, PathwaySet, RefID,
  TestClassIter, SampleIter)
}
\arguments{
\item{ExpMat_Test}{Expression matrix of test samples}

\item{ExpMat_Ref}{Expression matrix of reference samples}

\item{GeneVec}{Vector of gene names}

\item{PathwaySet}{List of pathways containing gene annotations for each pathways}

\item{RefID}{Class of the reference set}

\item{TestClassIter}{Class of the test set 
(if it is the same as reference set, the target test sample will be removed fro the reference set)}

\item{SampleIter}{Target test sample in ExpMat_Testto be used for comparison with ExpMat_Ref}
}
\value{
List of similarities between the target sample and the expression matrix of reference samples
}
\description{
Similarities_Wrapper is wrapper to identify similarities between the expression of genes in target sample
and the reference expression matrix
}
