% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSummary_2Class.R
\name{SimSummary_2Class}
\alias{SimSummary_2Class}
\title{SimSummary_2Class is a function to calculating similarity between two set of samples}
\usage{
SimSummary_2Class(SimMat1, SimMat2)
}
\arguments{
\item{SimMat1}{Matrix of similarity of the target samples with the 1st reference matrix}

\item{SimMat2}{Matrix of similarity of the target samples with the 2nd reference matrix}
}
\value{
Matrix of similarities of samples
}
\description{
SimSummary_2Class is a function to calculating similarity between two set of samples
}
