% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subopt_avian.R
\docType{data}
\name{subopt_avian}
\alias{subopt_avian}
\title{Subset of Avian Suboptimal Choice Data from Dunn et al. (2024)}
\format{
A data frame with 128 rows and 32 columns:
\describe{
\item{row}{Original row index in the full dataset \link{subopt_full}, retained for traceability and reproducibility.}
\item{study}{Short identifier for the study from which the data and parameters were extracted.}
\item{year}{Year the study was published.}
\item{species}{Species tested in the experiment (e.g., pigeon or starling).}
\item{exp}{Experiment number as reported in the original study. Values joined by an '&' indicate that data were aggregated across experiments.}
\item{condition}{Short label used to identify conditions within each study, aiding in dataset assembly.}
\item{n}{Number of subjects in the condition.}
\item{cp}{Observed choice proportion reported by the study.}
\item{il_dur_a, il_dur_b}{Average programmed initial link (IL) durations for alternatives A and B, respectively. FR1 schedules are assigned a duration of 1; other fixed-ratio schedules are given a duration of \eqn{1 + (\text{ratio}/\text{run rate})}.}
\item{tl_dur_a1, tl_dur_a2, tl_dur_b1, tl_dur_b2}{Average programmed durations of terminal links. For example, \code{tl_dur_a1} corresponds to the first terminal link on alternative A.}
\item{tl_p_a1, tl_p_a2, tl_p_b1, tl_p_b2}{Probability of entering each respective terminal link.}
\item{tr_p_a1, tr_p_a2, tr_p_b1, tr_p_b2}{Probability of terminal (i.e., primary) reinforcement following entry into each respective terminal link.}
\item{il_sched_a, il_sched_b}{Schedule type used for the initial links (e.g., VI, FR).}
\item{tl_sched_a1, tl_sched_a2, tl_sched_b1, tl_sched_b2}{Schedule type used for each terminal link.}
\item{forced_exposure}{Ordinal variable describing the amount of forced exposure (FE) trials used in the study. See \code{Details}.}
\item{DOI}{Digital Object Identifier (DOI) for the original study.}
\item{ref}{Full bibliographic reference for the study.}
\item{data_version}{A character column specifying the dataset version
based on the date it was last manually curated or corrected
(e.g., "2025-07-20"). Used for tracking updates to the compiled data.}
}
}
\source{
Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2024).
Suboptimal choice: A review and quantification of the signal for good news
(SiGN) model. \emph{Psychological Review}. \emph{131}(1), 58-78.
\url{https://doi.org/10.1037/rev0000416}
}
\usage{
subopt_avian
}
\description{
This dataset contains compiled avian suboptimal choice data (pigeons and starlings)
from multiple published studies that meet the inclusion criteria outlined in
Dunn et al. (2024). The subset was selected for model evaluation.
}
\details{
\emph{No FE:} Did not include any FE trials;
Only 1 FE session at the start of the condition;
FE trials for only some subjects.

\emph{Few FE:} 4 - 20 FE trials at the start of each session;
6 - 12 sessions of FE at the start of the first condition in the experiment;
2 sessions at the start of each condition.

\emph{Lots of FE:} FE trial following every suboptimal blackout; At least 50\% of all trials were FE.
}
\keyword{datasets}
