% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Stats.R
\name{Survival_Stats}
\alias{Survival_Stats}
\title{Survival_Stats is a function for building cox model using all the features and each feature as a separate model}
\usage{
Survival_Stats(ScoreMat, TimeVec, EventVec)
}
\arguments{
\item{ScoreMat}{Matrix of feature values used for survival predition}

\item{TimeVec}{Vectore of time to death of samples (patients)}

\item{EventVec}{Vector of events for the samples (patients) as being dead or censored}
}
\value{
A list containing summary of a cox model using all of the features and separate cox models for each feature
}
\description{
Survival_Stats is a function for building cox model using all the features and each feature as a separate model
}
