% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalStat_PostProcess.R
\name{SurvivalStat_PostProcess}
\alias{SurvivalStat_PostProcess}
\title{SurvivalStat_PostProcess is a function to Extract summary statistics of the built cox model}
\usage{
SurvivalStat_PostProcess(StatList)
}
\arguments{
\item{StatList}{Summary lists of the cox models built using all the}
}
\value{
A list including Cindex, Cindex_std and LogTest_pval
}
\description{
SurvivalStat_PostProcess is a function to Extract summary statistics of the built cox model
}
