% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSC.R
\name{TSC}
\alias{TSC}
\title{TSC is a function to calculate transcriprtional similarity coefficient between two biological pathways}
\usage{
TSC(PathwayExp1, PathwayExp2)
}
\arguments{
\item{PathwayExp1}{Expression matrix of genes within the chosen pathway in the 1st set of samples}

\item{PathwayExp2}{Expression matrix of genes within the chosen pathway in the 2nd set of samples}
}
\value{
Transcriptional similarity coefficient
}
\description{
TSC is a function to calculate transcriprtional similarity coefficient between two biological pathways
}
\examples{
Pathway1_ExpMat <- matrix(runif(100,0,10), ncol = 10)
Pathway2_ExpMat <- matrix(runif(100,0,10), ncol = 10)
TSC(Pathway1_ExpMat, Pathway2_ExpMat)
}
