% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_sched_input.R
\name{valid_sched_input}
\alias{valid_sched_input}
\title{Validate Schedule Type Input}
\usage{
valid_sched_input(x)
}
\arguments{
\item{x}{A list of 2 character vectors, where each vector contains schedule
labels. See \code{\link[=choice_params]{choice_params()}} for additional details about inputting
schedule types.}
}
\value{
A logical value: \code{TRUE} if all pairs contain valid and matching
schedule types.
}
\description{
Check whether pairs of schedule types input are valid. Specifically, each
pair must be either "VI" "VI" or "FR", "FR".
}
\examples{
# Valid
valid_sched_input(list(c("VI", "VI", "FR"), c("VI", "VI", "FR")))
# Invalid
valid_sched_input(list(c("FR", "VI", "FR"), c("VI", "VI", "FR")))
}
\seealso{
\code{\link[=choice_params]{choice_params()}}
}
