% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_params.R
\name{choice_params}
\alias{choice_params}
\title{Generate SiGN Model Parameter List}
\usage{
choice_params(
 profile = c("zentall", "kendall", "fantino"),
 il_dur_a = NULL, il_dur_b = NULL,
 tl_dur_a1 = NULL, tl_dur_a2 = NULL,
 tl_dur_b1 = NULL, tl_dur_b2 = NULL,
 tl_p_a1 = NULL, tl_p_a2 = NULL,
 tl_p_b1 = NULL, tl_p_b2 = NULL,
 tr_p_a1 = NULL, tr_p_a2 = NULL,
 tr_p_b1 = NULL, tr_p_b2 = NULL,
 il_sched_a = NULL, il_sched_b = NULL,
 s_delta = 1,
 beta_log = 10,
 beta_toggle = TRUE,
 display_params = FALSE
 )
}
\arguments{
\item{profile}{A character string specifying a default parameter profile. Must be one of
\code{"zentall"}, \code{"kendall"}, or \code{"fantino"}. See details.}

\item{il_dur_a, il_dur_b}{Numeric vector of initial link durations for
alternatives A and B respectively.}

\item{tl_dur_a1, tl_dur_a2}{Numeric vector of durations for two
terminal links on alternative A.}

\item{tl_dur_b1, tl_dur_b2}{Numeric vector of durations for two
terminal links on alternative B.}

\item{tl_p_a1, tl_p_a2}{Numeric vector of entry probabilities for two
terminal links on alternative A.}

\item{tl_p_b1, tl_p_b2}{Numeric vector of entry probabilities for two
terminal links on alternative B.}

\item{tr_p_a1, tr_p_a2}{Numeric vector of terminal reinforcement probabilities for two
terminal links on alternative A.}

\item{tr_p_b1, tr_p_b2}{Numeric vector of terminal reinforcement probabilities for two
terminal links on alternative B.}

\item{il_sched_a, il_sched_b}{Schedule type for each initial link. Must be
either \code{"VI"} (variable-interval) or \code{"FR"} (fixed-ratio). Case-insensitive.}

\item{s_delta}{Time required to perceive a stimulus that signals the
absence of terminal reinforcement. Default is 1. See also \code{\link[=s_delta]{s_delta()}}.}

\item{beta_log}{A positive real number specifying the base of
the logarithm used for β. Defaults to 10.}

\item{beta_toggle}{Logical defaulting to \code{TRUE}. Permits the model to adjust
the balance of conditional and primary reinforcement for signalled alternatives according to Equation 6 of Dunn et al. (2024). See also \code{\link[=beta_sig]{beta_sig()}}.}

\item{display_params}{Logical. If \code{TRUE}, prints the parameter list as a
data frame.}
}
\value{
A named list of validated and possibly recycled parameters suitable
for input into the \code{SiGN()} function.
}
\description{
A helper function that constructs a named list of parameters for behavioural
choice predictions for the function \code{\link[=SiGN]{SiGN()}}. Users can either select a
predefined profile or specify custom values. The function includes built-in
validation and recycling of arguments when lengths differ.
}
\details{
\emph{Default Profiles:}

Three default profiles are available to streamline argument selection:
\code{"zentall"} (based on Stagner & Zentall, 2010), \code{"kendall"} (based on
Kendall, 1985), and \code{"fantino"} (based on Fantino, 1969). Notably, the "
fantino" profile reflects a single terminal link procedure, consistent with
the design used in Fantino’s original study.

Default profile values can be overridden by specifying the corresponding
arguments directly.

If argument lengths differ, values are recycled to the maximum length,
with a message.

\emph{Temporal Units:}

The SiGN model is temporally relative, meaning it allows initial and
terminal link durations to be treated as any unit of time (e.g., seconds,
minutes, hours, etc.). However, units need to be
consistent across all link durations.

\emph{s_delta}

Conventionally, concurrent-chains paradigms have measured time in seconds
with very easily discriminable stimuli. The default value of \code{s_delta}
reflects this convention. However, if shorter or longer temporal units are
used, or the stimuli takes considerable time to be discriminated, it may be
prudent to adjust \code{s_delta} accordingly.

\emph{Signal Discrimination:}

A choice alternative is treated by the \code{\link[=SiGN]{SiGN()}} function as signalled if
both terminal link durations are non-zero and either the durations or
the reinforcement probabilities differ between the terminal links.

This definition assumes that differences in duration or reinforcement
probability imply discriminability between the stimuli associated with
each terminal link. For example, if one terminal link lasts 10 seconds
and the other 20 seconds, and both have a reinforcement probability of 1,
it is assumed that the organism can fully discriminate between them—e.g.,
the terminal links may be visually distinct (such as different colours)
and thus are recognisable from the moment of onset.

The SiGN model assumes that the (behavioural) function of terminal link
stimuli remains fixed across the entire duration of the terminal link—that
is, the probability of terminal reinforcement does not change
partway through the terminal link. Procedures that depart from this
assumption may not be appropriately handled by the model.

\emph{Schedule Selection:}

Setting the initial link schedules as \code{"FR"} is most suitable for cases
involving an FR 1 schedule or when a single timer is employed in the
initial links of long VI schedules. This is because the \code{"FR"} setting does
not take into account the switching behaviour present with concurrent
schedules using independent timers. For example, if a single timer is used
for two concurrent VI 30 schedules, setting the Initial Link Schedule as
an FR is preferable because the time spent in the initial links is
controlled by one timer, not two independent timers operating concurrently.
However, if independent timers are used for each initial link, the model
requires this setting to be VI .

Note that the SiGN model does not directly compute predictions for ratio
schedules. Instead, it represents the ratio as a duration with an individual
reinforcement rate, rather than a common/shared one.
}
\examples{
# Use default Zentall profile
z <- choice_params("zentall", display_params = TRUE)

# Custom specification with minimal values
k <- choice_params(
  profile = "kendall",
  tr_p_b1 = c(1, 0.5), tr_p_b2 = c(1, 0.5),
  display_params = TRUE
)
}
\references{
Dunn, R. M., Pisklak, J. M., McDevitt, M. A., & Spetch, M. L. (2024).
Suboptimal choice: A review and quantification of the signal for good news
(SiGN) model. \emph{Psychological Review}. \emph{131}(1), 58-78.
\url{https://doi.org/10.1037/rev0000416}

Fantino, E. (1969). Choice and rate of reinforcement. \emph{Journal of the
Experimental Analysis of Behavior}, \emph{12}(5), 723–730.
\url{https://doi.org/10.1901/jeab.1969.12-723}

Kendall, S. B. (1985). A further study of choice and percentage
reinforcement. \emph{Behavioural Processes}, \emph{10}(4), 399–413.
\url{https://doi.org/10.1016/0376-6357(85)90040-3}

Stagner, J. P., & Zentall, T. R. (2010). Suboptimal choice behavior by
pigeons. \emph{Psychonomic Bulletin & Review}, \emph{17}(3), 412–416.
\url{https://doi.org/10.3758/PBR.17.3.412}
}
\seealso{
\code{\link[=dur_entry_conflict]{dur_entry_conflict()}}, \code{\link[=entry_p_sums]{entry_p_sums()}}, \code{\link[=valid_sched_input]{valid_sched_input()}}, \code{\link[=recycle_list]{recycle_list()}}, \code{\link[=eq_arg_n]{eq_arg_n()}}]
}
