% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_check.R
\name{sig_check}
\alias{sig_check}
\title{Determine Whether a Terminal Link Alternative Is Signalled}
\usage{
sig_check(dur1, dur2, rein_p1, rein_p2)
}
\arguments{
\item{dur1}{A numeric vector giving the durations of an alternative's first
TL.}

\item{dur2}{A numeric vector giving the durations of an alternative's second
TL.}

\item{rein_p1}{A numeric vector giving the terminal reinforcement
probabilities for an alternative's first TL.}

\item{rein_p2}{A numeric vector giving the terminal reinforcement
probabilities for an alternative's second TL.}
}
\value{
A logical vector indicating whether the choice alternative can be
considered signalled.
}
\description{
Evaluates whether an alternative can be considered "signalled" on the basis
of its two terminal link (TL) durations and reinforcement probabilities.
}
\details{
The durations and probabilities input should correspond to only one choice
alternative.

A choice alternative is treated as signalled if both terminal link (TL)
durations are non-zero and either the durations or the terminal
reinforcement probabilities differ between the two TLs.

This definition assumes that differences in duration or reinforcement
probability imply discriminability between the stimuli associated with each
TL. For example, if one TL lasts 10 seconds and the other 20 seconds, and
both have a reinforcement probability of 1, it is assumed that the
organism can fully discriminate between them—e.g., the TLs may be visually
distinct (such as different colours) and thus are recognisable from the
moment of onset. See also \code{\link[=choice_params]{choice_params()}}
}
\examples{
sig_check(10, 10, 0.5, 0.5)  # FALSE — same duration and probability
sig_check(10, 10, 0.5, 0.0)  # TRUE  — same duration, different probability
sig_check(10, 5, 0.5, 0.5)   # TRUE  — different durations
sig_check(0, 10, 0.5, 0.5)   # FALSE — one duration is zero
}
